fix: refactor register page to pages dir
