+++ src/app/app.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import {RegisterComponent} from "./components/register/register.component";
+import {RegisterComponent} from "./pages/register/register.component";
 import { LoginComponent } from './pages/login/login.component';
 import {ProfileComponent} from './pages/profile/profile.component';
