+++ src/app/pages/register/form-validator.ts
